pyproject.toml
[project]
name = "company-config-loader"
version = "0.1.0"
description = "Strict YAML config loader with deep-merge for base/env/local files"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["PyYAML>=6.0"]

[build-system]
requires = ["setuptools>=68", "wheel", "build"]
build-backend = "setuptools.build_meta"
